requirements.txt
openai>=1.5.0
python-dotenv==1.0.0
pyaudio==0.2.14
SpeechRecognition
pvporcupine
numpy
pytz
anthropic
pydub
tzdata>=2023.3
python-dateutil>=2.8.2
requests>=2.31.0
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
zeep>=4.2.1
mcp>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
pyyaml>=6.0.0